refactor(steam): split proton.rs into directory module
Splits the 1090-line `steam/proton.rs` into a `steam/proton/` directory
module with logical submodules (types, resolution, discovery, matching,
flatpak, util) plus a `tests/` subdirectory grouped by cluster.

Public API and re-exports preserved; cargo test -p crosshook-core,
lint.sh, and check-host-gateway.sh all green.

Part of #290
